requirements.txt
absl-py==2.3.1
accelerate==1.13.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==26.1.0
basicsr==1.4.2
bitsandbytes==0.49.2
cachetools==7.0.2
causal_conv1d==1.6.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.3
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
datasets==4.8.4
dill==0.4.1
distlib==0.4.0
einops==0.8.2
exceptiongroup==1.3.1
facexlib==0.3.0
filelock==3.25.2
filterpy==1.4.5
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gfpgan==1.3.8
google-auth==2.48.0
google-auth-oauthlib==1.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
humanfriendly==10.0
identify==2.6.18
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lit==18.1.8
llvmlite==0.45.1
lmdb==1.7.5
mamba_ssm==2.3.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.4.2
ninja==1.13.0
nodeenv==1.10.0
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnx==1.17.0
onnx-ir==0.1.12
onnxruntime==1.13.1
onnxscript==0.2.5
openai-clip==1.0.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.5.1
propcache==0.4.1
protobuf==6.32.1
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiqa==0.1.15.post2
pyparsing==3.2.5
pytest==9.0.3
python-dateutil==2.9.0.post0
python-discovery==1.2.2
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
ruff==0.15.10
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.1
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tifffile==2025.5.10
timm==1.0.25
tokenizers==0.22.2
tomli==2.3.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==5.2.0
triton==3.5.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==21.2.1
wcwidth==0.6.0
Werkzeug==3.1.3
xxhash==3.6.0
yapf==0.43.0
yarl==1.23.0
zipp==3.23.0
